Added sign-in and sign-up page functionality using Clerk components. Added Home and About Us pages to allowed pages when not logged in. Added different signup duck image than login.
