+++ src/app/sign-up/layout.jsx
@@ -1,0 +1,9 @@
+import Topbar from "../ui/topbar";
+
+export default function Layout({ children }) {
+  return (
+    <div>
+      <Topbar>{children}</Topbar>
+    </div>
+  );
+}
